featurize/pyproject.toml
[project]
name = "trak-manual"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.35.88",
    "cloudpathlib>=0.20.0",
    "hydra-core>=1.3.2",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "numba>=0.58.1",
    "numpy>=1.24.4",
    "open-clip-torch>=2.30.0",
    "pandas>=2.0.3",
    "pyarrow>=17.0.0",
    "rich>=13.9.4",
    "s3fs>=2024.10.0",
    "setuptools>=75.3.0",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "tqdm>=4.67.1",
    "training",
    "traker[fast]>=0.3.2",
    "webdataset>=0.2.100",
    "zarr>=2.16.1",
]

[tool.uv.sources]
training = { path = "../datacomp/training" }
